requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.5.0
python-dotenv>=1.0.0

chromadb
sentence-transformers

huggingface-hub>=0.19.4
transformers>=4.35.2
langchain-community
tokenizers>=0.15.0

numpy
pandas
tqdm>=4.66.1
json-repair>=0.1.5

aiofiles>=23.2.1
httpx>=0.25.1
python-multipart>=0.0.6

python-json-logger>=2.0.7

streamlit
google-generativeai
langchain-core
langchain-google-genai
langchain-huggingface
langchain-chroma
requests
plotly
spacy
transformers
torch
langchain-community

# Pinecone dependencies
pinecone-client>=2.2.4
langchain-pinecone>=0.0.1